@@ -3,7 +3,7 @@ description: A flutter project which add scroll direction to ListWheelScrollView
 version: 0.0.2
 repository: https://github.com/jascodes/list_wheel_scroll_view_x
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0 <2.0.0"
 
 dependencies:
