@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class ColoredContainer extends StatelessWidget {
   final Color color;
-  const ColoredContainer(this.color, {Key key}) : super(key: key);
+  const ColoredContainer(this.color, {Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
